Fix small bugs 🐛
