+++ src/screens/NewTaskScreen/components/AttachmentButton.tsx
@@ -4,7 +4,8 @@ import {
   StyleSheet,
   View,
   Modal,
-  Dimensions
+  Dimensions,
+  Alert
 } from 'react-native';
 import * as ImagePicker from 'expo-image-picker';
 import Constants from 'expo-constants';
@@ -25,7 +26,7 @@ const AttachmentButton: React.FC<Props> = ({ onGetImages }) => {
         Permissions.CAMERA
       );
       if (status !== 'granted') {
-        alert('Sorry, we need camera roll permissions to make this work!');
+        Alert.alert('Sorry, we need camera roll permissions to make this work!');
       }
     }
   };
